requirements.txt
tensorflow>=2.10.0
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.1.0
matplotlib>=3.6.0
seaborn>=0.12.0
Pillow>=9.0.0
opencv-python>=4.6.0
streamlit>=1.22.0
